Reverted heatmiser tool to the same file
